Update main.bicep
Key Changes
Current Time Variable: A currentTime variable is defined using the utcNow() function, which generates a timestamp at the time of deployment. This is formatted as 'yyyyMMddHHmmss' to avoid any characters that are not allowed in Azure resource names.
Module Name and Function Name: Both the name parameter of the module and the functionname parameter passed to the function module include currentTime along with uniqueSuffix and amlworkspace. This approach drastically reduces the chance of name conflict